Modified initial guess testing
Added simple algorithm to handle obstacle collision in the initial guess function. If any of the predicted states ends up colliding with an obstacle, it will change the steering angle slightly and try again. No detection for input constraints in the initial guess is implemented yet.

Simplified the constructors for different kinds of controlllers.