Additional controller refactoring
Removed another unnecessary optimization variable state (z_0, which is
constant).

Found some unnecessary inequality and equality constraints, which I
suspect caused some estimated matrix inside the solver to be singular
which terminated the solved prematurely. The solver now can run for a
couple of time steps, but the solution quickly diverges and escapes the
feasible space which again causes the solver to terminate prematurely.