Improved initial guess function
Initial guesser now works based on the last optimal value of the solver.
It takes the previous values and tries to apply the same control signal
again and sees if it ends up somewhere legal. If not, it iterates
backwards and steers away from the obstacle until the new point ends up
outside it.

Ran sim on N=7 and added.